requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
openai
python-dotenv
pypdf